chore: extract db package
