+++ apps/server/package.json
@@ -9,6 +9,7 @@
     },
   "dependencies": {
     "@aws-sdk/client-sqs": "^3.678.0",
+    "@gbt/db": "*",
     "@google-cloud/translate": "8.5.0",
     "@gracious.tech/fetch-client": "0.8.9",
     "@next/env": "14.2.12",
@@ -19,7 +20,6 @@
     "@opentelemetry/instrumentation": "0.56.0",
     "@opentelemetry/instrumentation-aws-sdk": "0.48.0",
     "@opentelemetry/instrumentation-http": "0.56.0",
-    "@opentelemetry/instrumentation-pg": "0.49.0",
     "@opentelemetry/propagator-aws-xray": "1.26.0",
     "@opentelemetry/resources": "1.29.0",
     "@opentelemetry/sdk-logs": "0.56.0",
@@ -36,8 +36,6 @@
     "nodemailer": "6.9.15",
     "npm": "10.8.3",
     "oslo": "1.2.1",
-    "pg": "8.12.0",
-    "pg-cursor": "2.12.1",
     "react": "^18",
     "react-dom": "^18",
     "swr": "2.2.5",
@@ -61,8 +59,6 @@
     "@types/lodash": "4.17.7",
     "@types/node": "^20",
     "@types/nodemailer": "6.4.15",
-    "@types/pg": "8.11.6",
-    "@types/pg-cursor": "2.7.2",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "chart.js": "4.4.4",
