+++ packages/db/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "@gbt/db",
+    "type": "module",
+    "scripts": {
+        "dev": "tsc --watch",
+        "build": "tsc",
+        "check-types": "tsc --noEmit"
+    },
+    "exports": {
+        "./query": {
+            "types": "./src/query.ts",
+            "default": "./dist/query.js"
+        }
+    },
+    "dependencies": {
+        "@opentelemetry/instrumentation-pg": "0.49.0",
+        "pg": "8.12.0",
+        "pg-cursor": "2.12.1"
+    },
+    "devDependencies": {
+        "@gbt/typescript-config": "*",
+        "@types/pg": "8.11.6",
+        "@types/pg-cursor": "2.7.2",
+        "typescript": "^5"
+    }
+}
